add: 詳細ページに切り替わるように
