+++ src/App.css
@@ -1,5 +1,7 @@
 .App {
   text-align: center;
+  background-color: #FFFFF0; /* こちらを追加 */
+
 }
 
 .App-logo {
@@ -15,13 +17,14 @@
 
 .App-header {
   background-color: #FFFFF0;
-  min-height: 100vh;
+  min-height: 10vh;
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: center;
+  justify-content: flex-start; 
   font-size: calc(10px + 2vmin);
   color: black;
+  padding-top: 10px;
 }
 
 .App-link {
@@ -38,9 +41,7 @@
 }
 
 .pokemon-grid {
-  /* ポケモンを3匹ずつ、3行で表示するためのスタイル */
   display: grid;
-  grid-template-columns: repeat(3, 1fr); /* 3列のグリッド */
-  gap: 30px; /* グリッドアイテム間の間隔 */
-  /* 他のスタイルプロパティを追加 */
+  grid-template-columns: repeat(3, 1fr);
+  gap: 30px;
 }
