+++ src/components/Detail.tsx
@@ -1,11 +1,14 @@
-import React from 'react'
+import React from 'react';
+import { useParams } from 'react-router-dom'; // useParamsのインポート
 
 function Detail() {
+    const { pokemonName } = useParams<{ pokemonName: string }>(); // パラメータからポケモンの名前を取得
+
     return (
         <div>
-            aboutページ
+            {pokemonName}の詳細ページ
         </div>
     )
 }
 
-export default Detail
+export default Detail;
